dbgproto.h and dbgdefs.h moved to PRinternal
