+++ src/rmon/rmonmain.c
@@ -10,7 +10,7 @@
 
 #ifndef _FINALROM
 
-#include "dbgproto.h"
+#include "PRinternal/dbgproto.h"
 #include "PR/os_internal.h"
 #include "PRinternal/rmonint.h"
 #include "PR/rcp.h"
